api/pyproject.toml
[project]
name = "google-ad-automation-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9",
]
